Add getComputerChoice() random choice function
Why:
This function adds the capability for the computer player to generate a
random choice (rock, paper or scissors) and output the choice for use
elsewhere in the program.

How:
- getComputerChoice() function created within app.js with internal
variable declarations. RETURN value from function is 'choice' and is
text, either 'rock', 'paper' or 'scissors'

- Made use of Math.floor(Math.random() * array length) function to
generate a random number between 0 and 2. And then applied this to the
possibleChoices array I had created to find the computer choice - this
value would then be assigned to the choice variable and returned from
the function.

Current TODO: List / Bug list
TODO: Write playRound() that plays a single round of Rock Paper Scissors

TODO: Write game() to play 5 rounds and report winner