Add game() to play 5 rounds and report winner
Why:
This commit adds the game() function which plays first to 5 rounds and
then reports the winner.
* All functionality is only currently within Console*

How:
- Changed playRound() function slightly, adding a playerScore and
computerScore increment on each feedback. This feedback happens before
the return from the function.

- Implemented the game() function which takes the user input via a
prompt. This is currently very basic and provides very little feedback
currently other than through the Console. Where the result of the playRound
is shown.

- Currently the game() function does a comparison of the scores at the
end of the loop (which only runs while both playerScore and
computerScore < 5) via an IF statement. And then returns a statement of
who the winner is. This could be developed further if need be later.

Current TODO: List / Bug list
TODO: Add GUI